pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "ssi"
authors = [
{ name = "Anil Keshwani", email = "anil.keshwani.academic@gmail.com" },
]
description = "Research implementation to investigate methods of integrating the speech modality into pre-trained language models"
requires-python = "==3.10.6"
license = { file = "LICENSE" }
dynamic = ["readme", "version"]
dependencies = [
    "evaluate>=0.4.3",
    "huggingface_hub==0.25.2",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "inflect==7.3.1",
    "jiwer>=3.1.0",
    "matplotlib==3.10.6",
    "omegaconf==2.3.0",
    "psutil==6.0.0",
    "pyarrow==17.0.0",
    "PyYAML==6.0.2",
    "regex==2024.9.11",
    "scikit-learn==1.7.2",
    "seaborn==0.13.2",
    "sentencepiece==0.2.0",
    "six==1.16.0",
    "tiktoken==0.8.0",
    "torch==2.5.1",
    "torchao==0.5.0",
    "torchaudio>=2.2.2",
    "torchtune==0.5.0",  # keep strict
    "torchvision==0.20.1",
    "tqdm==4.66.5",
    "vllm>=0.7.3; sys_platform == 'linux'",
    "wandb>=0.20.1",
    "whisper_normalizer>=0.1.11",
]

[tool.setuptools]
packages = ["ssi"] # Sardine Speech Integration

[tool.setuptools_scm]
version_file = "ssi/_version.py"

[tool.setuptools.dynamic]
readme = { file = ["README.md"] }

[project.optional-dependencies]
dev = [
"black==24.4.2",
"flake8",
"isort==5.13.2",
# test runner
"pytest>=8.0.0",
# "mypy",
"pipreqs==0.5.0",
"pre-commit==3.8.0",
"shfmt-py==3.7.0.1",
"usort==1.0.8",
]

[tool.uv]
environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
force_alphabetical_sort_within_sections = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: requires CUDA GPU resources",
    "regression: long-running regression tests",
    "heavy: expensive end-to-end runs",
    "tier_smoke: short smoke-profile tests",
    "tier_full: full-profile tests",
    "approach_cpt: CPT training path",
    "approach_sft: SFT training path",
    "approach_cpt_sft: CPT then SFT chain",
    "tokenizer_hubert: HuBERT dataset/tokenizer path",
    "tokenizer_speechtokenizer: SpeechTokenizer dataset/tokenizer path",
    "tokenizer_mimi: Mimi dataset/tokenizer path",
    "tokenizer_focalcodec: FocalCodec dataset/tokenizer path",
]
